requirements.txt
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.40
botocore==1.40.40
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
coverage==7.10.6
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-silk==5.4.3
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.28.0
gprof2dot==2025.4.14
gunicorn==23.0.0
hiredis==3.2.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
